requirements.txt
# Core AI and Image Processing
replicate==0.9.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0

# Configuration and Data Handling
PyYAML>=6.0
requests>=2.31.0
click>=8.1.0
tqdm>=4.66.0
rich>=13.5.0
typer>=0.9.0
python-dotenv>=1.0.0

# PyTorch and Deep Learning (compatible versions with basicsr)
torch>=1.13.0,<2.1.0
torchvision>=0.14.0,<0.16.0
torchtext>=0.6.0
torchaudio>=0.13.0,<2.1.0

# Stable Diffusion and ControlNet
diffusers>=0.21.0
transformers>=4.33.0
accelerate>=0.21.0
xformers>=0.0.20

# Real-ESRGAN and Upscaling
git+https://github.com/xinntao/Real-ESRGAN.git
basicsr==1.4.2
gfpgan==1.3.8

# Image Enhancement and Processing
scikit-image>=0.21.0
imageio>=2.37.0
colour-science>=0.4.3

# LoRA and Model Enhancement
peft>=0.4.0
safetensors>=0.3.0

# Advanced Image Processing
wand>=0.6.11
Pillow-SIMD>=9.0.0

# Color Science and Professional Processing
opencv-contrib-python>=4.8.0

# Post-processing and Quality Verification
psutil>=5.9.0
colorama>=0.4.6
lpips>=0.1.4

# Face Analysis and Quality Control
face-recognition>=1.3.0
dlib>=19.24.0

# Development and Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
